+++ tests/Unit/RestApi/Applet/Command/AppletCommandManagement.spec.ts
@@ -1,8 +1,11 @@
 import * as should from 'should';
-import * as nock from "nock";
-import { nockOpts } from "../../helper";
-import AppletCommandManagement from "../../../../../src/RestApi/Applet/Command/AppletCommandManagement";
-import IAppletCommand, { IAppletCommandSendable } from "../../../../../src/RestApi/Applet/Command/IAppletCommand";
+import * as nock from 'nock';
+
+import { getNockOpts } from '../../helper';
+import AppletCommandManagement from '../../../../../src/RestApi/Applet/Command/AppletCommandManagement';
+import IAppletCommand, { IAppletCommandSendable } from '../../../../../src/RestApi/Applet/Command/IAppletCommand';
+
+const nockOpts = getNockOpts({});
 
 describe('AppletCommandManagement', () => {