+++ tests/Unit/RestApi/Device/Firmware/DeviceFirmwareManagement.spec.ts
@@ -1,9 +1,12 @@
 import * as should from 'should';
-import * as nock from "nock";
-import { nockOpts } from "../../helper";
-import IDeviceChangeResponse from "../../../../../src/RestApi/Device/IDeviceChangeResponse";
-import DeviceFirmwareManagement from "../../../../../src/RestApi/Device/Firmware/DeviceFirmwareManagement";
-import IDeviceFirmware, { IDeviceFirmwareUpdatable } from "../../../../../src/RestApi/Device/Firmware/IDeviceFirmware";
+import * as nock from 'nock';
+
+import { getNockOpts } from '../../helper';
+import IDeviceChangeResponse from '../../../../../src/RestApi/Device/IDeviceChangeResponse';
+import DeviceFirmwareManagement from '../../../../../src/RestApi/Device/Firmware/DeviceFirmwareManagement';
+import IDeviceFirmware, { IDeviceFirmwareUpdatable } from '../../../../../src/RestApi/Device/Firmware/IDeviceFirmware';
+
+const nockOpts = getNockOpts({});
 
 describe('DeviceFirmwareManagement', () => {